// packages/tsts/src/ast/generated/schema.ts
export const NodeDefinitions = {
  "Token": {
    "extends": [
      "NodeBase"
    ],
    "typeParameters": [
      {
        "name": "TKind",
        "constraint": "TokenSyntaxKind",
        "default": "TokenSyntaxKind"
      }
    ],
    "members": [
      {
        "name": "Kind",
        "type": "TKind",
        "inherited": true
      }
    ],
    "arena": true,
    "instantiationAliases": {
      "EndOfFile": "EndOfFile",
      "DotToken": "DotToken",
      "DotDotDotToken": "DotDotDotToken",
      "QuestionToken": "QuestionToken",
      "ExclamationToken": "ExclamationToken",
      "ColonToken": "ColonToken",
      "EqualsToken": "EqualsToken",
      "AsteriskToken": "AsteriskToken",
      "EqualsGreaterThanToken": "EqualsGreaterThanToken",
      "PlusToken": "PlusToken",
      "MinusToken": "MinusToken",
      "QuestionDotToken": "QuestionDotToken",
      "AssertsKeyword": "AssertsKeyword",
      "AssertKeyword": "AssertKeyword",
      "AwaitKeyword": "AwaitKeyword",
      "CaseKeyword": "CaseKeyword",
      "AbstractKeyword": "AbstractKeyword",
      "AccessorKeyword": "AccessorKeyword",
      "AsyncKeyword": "AsyncKeyword",
      "ConstKeyword": "ConstKeyword",
      "DeclareKeyword": "DeclareKeyword",
      "DefaultKeyword": "DefaultKeyword",
      "ExportKeyword": "ExportKeyword",
      "InKeyword": "InKeyword",
      "PrivateKeyword": "PrivateKeyword",
      "ProtectedKeyword": "ProtectedKeyword",
      "PublicKeyword": "PublicKeyword",
      "ReadonlyKeyword": "ReadonlyKeyword",
      "OutKeyword": "OutKeyword",
      "OverrideKeyword": "OverrideKeyword",
      "StaticKeyword": "StaticKeyword",
      "BinaryOperatorToken": "BinaryOperator",
      "AssignmentOperatorToken": "AssignmentOperator"
    }
  },
  "Identifier": {
    "extends": [
      "PrimaryExpressionBase",
      "FlowNodeBase"
    ],
    "members": [
      {
        "name": "Text",
        "type": "string"
      }
    ],
    "arena": true
  },
// ...
  "ReturnStatement": {
    "extends": [
      "StatementBase",
      "CompositeBase"
    ],
    "members": [
      {
        "name": "Expression",
        "type": "Expression",
        "optional": true
      }
    ],
    "arena": true
  },
// ...
  "ParameterDeclaration": {
    "kind": "Parameter",
    "extends": [
      "NodeBase",
      "DeclarationBase",
      "ModifiersBase",
      "CompositeBase"
    ],
    "members": [
      {
        "name": "modifiers",
        "inherited": true
      },
      {
        "name": "DotDotDotToken",
        "type": "DotDotDotToken",
        "optional": true
      },
      {
        "name": "name",
        "type": "BindingName",
        "private": true
      },
      {
        "name": "QuestionToken",
        "type": "QuestionToken",
        "optional": true
      },
      {
        "name": "Type",
        "type": "TypeNode",
        "optional": true
      },
      {
        "name": "Initializer",
        "type": "Expression",
        "optional": true
      }
    ],
    "arena": true
  },
// ...
  "FunctionDeclaration": {
    "extends": [
      "DeclarationBase",
      "StatementBase",
      "ExportableBase",
      "ModifiersBase",
      "FunctionLikeWithBodyBase",
      "CompositeBase"
    ],
    "members": [
      {
        "name": "modifiers",
        "inherited": true
      },
      {
        "name": "AsteriskToken",
        "inherited": true
      },
      {
        "name": "name",
        "type": "Identifier",
        "private": true,
        "optional": true
      },
      {
        "name": "TypeParameters",
        "inherited": true
      },
      {
        "name": "Parameters",
        "inherited": true,
        "visit": "parameters"
      },
      {
        "name": "Type",
        "inherited": true
      },
      {
        "name": "FullSignature",
        "inherited": true,
        "noTS": true
      },
      {
        "name": "Body",
        "inherited": true,
        "type": "FunctionBody",
        "optional": true,
        "visit": "functionBody"
      },
      {
        "name": "ReturnFlowNode",
        "type": "*FlowNode",
        "goOnly": true
      }
    ],
    "arena": true
  },
// ...
  "SourceFile": {
    "extends": [
      "NodeBase",
      "DeclarationBase",
      "LocalsContainerBase",
      "CompositeBase"
    ],
    "members": [
      {
        "name": "Statements",
        "type": "Statement",
        "list": "NodeList",
        "visit": "topLevelStatements"
      },
      {
        "name": "EndOfFileToken",
        "type": "EndOfFile"
      }
    ],
    "handWritten": true
  },
// ...
} as const;

export const NodeAliases = {
// ...
  "BindingName": [
    "Identifier",
    "BindingPattern"
  ],
  "ModifierLike": [
    "Modifier",
    "Decorator"
  ],
  "JsxChild": [
    "JsxText",
    "JsxExpression",
    "JsxElement",
    "JsxSelfClosingElement",
    "JsxFragment"
  ],
  "JsxAttributeLike": [
    "JsxAttribute",
    "JsxSpreadAttribute"
  ],
  "JsxAttributeName": [
    "Identifier",
    "JsxNamespacedName"
  ],
  "JsxAttributeValue": [
    "StringLiteral",
    "JsxExpression",
    "JsxElement",
    "JsxSelfClosingElement",
    "JsxFragment"
  ],
  "JsxTagNameExpression": [
    "Identifier",
    "KeywordExpression",
    "PropertyAccessExpression",
    "JsxNamespacedName"
  ],
  "ClassLikeDeclaration": [
    "ClassDeclaration",
    "ClassExpression"
  ],
  "AccessorDeclaration": [
    "GetAccessorDeclaration",
    "SetAccessorDeclaration"
  ],
  "LiteralLikeNode": [
    "StringLiteral",
    "NumericLiteral",
    "BigIntLiteral",
    "RegularExpressionLiteral",
    "TemplateLiteralLikeNode",
    "JsxText"
  ],
  "LiteralExpression": [
    "StringLiteral",
    "NumericLiteral",
    "BigIntLiteral",
    "RegularExpressionLiteral",
    "NoSubstitutionTemplateLiteral"
  ],
  "UnionOrIntersectionTypeNode": [
    "UnionTypeNode",
    "IntersectionTypeNode"
  ],
  "TemplateLiteralLikeNode": [
    "PseudoLiteralSyntaxKind"
  ],
  "TemplateMiddleOrTail": [
    "TemplateMiddle",
    "TemplateTail"
  ],
  "TemplateLiteral": [
    "TemplateExpression",
    "NoSubstitutionTemplateLiteral"
  ],
  "TypePredicateParameterName": [
    "Identifier",
    "ThisTypeNode"
  ],
  "ImportAttributeName": [
    "Identifier",
    "StringLiteral"
  ],
  "LeftHandSideExpression": {
    "base": "LeftHandSideExpressionBase"
  },
// ...
} as const;

export const ListAliases = {
  "StatementList": "Statement",
  "CaseClausesList": "CaseOrDefaultClause",
  "VariableDeclarationNodeList": "VariableDeclaration",
  "BindingElementList": "BindingElement",
  "TypeParameterList": "TypeParameterDeclaration",
  "ParameterList": "ParameterDeclaration",
  "HeritageClauseList": "HeritageClause",
  "ClassElementList": "ClassElement",
  "TypeElementList": "TypeElement",
  "ExpressionWithTypeArgumentsList": "ExpressionWithTypeArguments",
  "EnumMemberList": "EnumMember",
  "ImportSpecifierList": "ImportSpecifier",
  "ExportSpecifierList": "ExportSpecifier",
  "TypeArgumentList": "TypeNode",
  "ArgumentList": "Expression",
  "TemplateSpanList": "TemplateSpan",
  "ElementList": "Expression",
  "PropertyDefinitionList": "ObjectLiteralElement",
  "TypeList": "TypeNode",
  "ImportAttributeList": "ImportAttribute",
  "TemplateLiteralTypeSpanList": "TemplateLiteralTypeSpan",
  "JsxChildList": "JsxChild",
  "JsxAttributeList": "JsxAttributeLike"
} as const;
// ...
export const NodeKindByName = {
// ...
  "SourceFile": "SourceFile",
  "ModuleDeclaration": "ModuleDeclaration",
  "ImportEqualsDeclaration": "ImportEqualsDeclaration",
  "ExportDeclaration": "ExportDeclaration",
  "ImportTypeNode": "ImportType",
  "ImportClause": "ImportClause",
  "ImportSpecifier": "ImportSpecifier",
  "JSDocText": "JSDocText",
  "JSDocLink": "JSDocLink",
  "JSDocLinkPlain": "JSDocLinkPlain",
  "JSDocLinkCode": "JSDocLinkCode",
  "TypeParameterDeclaration": "TypeParameter",
  "SyntheticReferenceExpression": "SyntheticReferenceExpression",
  "JSDocTypeLiteral": "JSDocTypeLiteral",
  "JSDocParameterOrPropertyTag": "JSDocParameterOrPropertyTag"
} as const;
// ...
export const NodeExtendsByName = {
// ...
  "SourceFile": [
    "NodeBase",
    "DeclarationBase",
    "LocalsContainerBase",
    "CompositeBase"
  ],
  "ModuleDeclaration": [
    "DeclarationBase",
    "StatementBase",
    "ExportableBase",
    "ModifiersBase",
    "LocalsContainerBase",
    "BodyBase",
    "CompositeBase"
  ],
  "ImportEqualsDeclaration": [
    "DeclarationBase",
    "StatementBase",
    "ExportableBase",
    "ModifiersBase",
    "CompositeBase"
  ],
  "ExportDeclaration": [
    "DeclarationBase",
    "StatementBase",
    "ModifiersBase",
    "CompositeBase"
  ],
  "ImportTypeNode": [
    "NodeWithTypeArgumentsBase"
  ],
  "ImportClause": [
    "NodeBase",
    "DeclarationBase",
    "ExportableBase",
    "CompositeBase"
  ],
  "ImportSpecifier": [
    "NodeBase",
    "DeclarationBase",
    "ExportableBase",
    "CompositeBase"
  ],
  "JSDocText": [
    "JSDocCommentBase"
  ],
  "JSDocLink": [
    "JSDocCommentBase"
  ],
  "JSDocLinkPlain": [
    "JSDocCommentBase"
  ],
  "JSDocLinkCode": [
    "JSDocCommentBase"
  ],
  "TypeParameterDeclaration": [
    "NodeBase",
    "DeclarationBase",
    "ModifiersBase",
    "TypeSyntaxBase"
  ],
  "SyntheticReferenceExpression": [
    "ExpressionBase"
  ],
  "JSDocTypeLiteral": [
    "JSDocTypeBase",
    "DeclarationBase"
  ],
  "JSDocParameterOrPropertyTag": [
    "JSDocTagBase"
  ]
} as const;